admin pages update
