+++ backend/analytics/models.py
@@ -1,5 +1,5 @@
 from django.db import models
-
+from django.utils import timezone
 
 class Campaign(models.Model):
     name = models.CharField(max_length=255)
@@ -34,7 +34,7 @@ class Click(models.Model):
     user_agent = models.CharField(max_length=255)
     os = models.CharField(max_length=50)
     browser = models.CharField(max_length=50)
-    click_time = models.DateTimeField(auto_now_add=True)
+    click_time = models.DateTimeField(default=timezone.now)
     landing_page_url = models.URLField()
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
